refactor: don't show console.error when testing
This `console.error` is only relevant if we're not yet dealing with this
error type, so moving it to the bottom of this function is more
appropriate

This also means we no longer have a loud console error when running the
tests. This console output made me think the tests weren't working at
first!