feat(recoveryMethods): return all info we store
The test restructuring wasn't entirely necessary, but I find it much
easier to understand this way.

Note also that the new `createNearAccount` function is not called in the
new tests, since there's no reason to actually create this account
on-chain, and doing so slows down the tests by ~100x. I extracted it
before I realized I wouldn't need it, and I think it gives increased
clarity to keep it separate like this.