feat(recoveryMethods): verify user owns account
Before displaying or deleting recovery methods, we need to ensure that
the person hitting the endpoint is the actual owner of the account. We
now do so by requiring requests to include a `securityCode` parameter,
which can currently be anything, and a `signature` derived from that
`securityCode`. We then check that the account being modified is, in
fact, owned by the person who signed.

This also changed the endpoint for deleting recovery methods and
switched it to use a `POST` request. While not necessary, this avoids a
lot of work in updating our dependencies to properly support sending a
JSON body along with a DELETE request (a valid thing to do by the HTTP
spec, but not well supported in HTTP libraries).