refactor: update deleteRecoveryMethod middlewares
* make `withAccount`, adding account to `ctx.account`
* rename `recoveryMethodUpdaters` to `recoveryMethodDeleters`
* re-order the two

Why `recoveryMethodDeleters` instead of using a switch statement? While
a switch statement might be more common and immediately readable,
there's a big benefit to using this `recoveryMethodDeleters` pattern: if
we want to add a new recovery method in the future, we add one key/value
to `recoveryMethodDeleters` and both `checkRecoveryMethod` and the route
itself both do the right thing. Using a switch statement would introduce
some easy-to-miss coupling between the two.